🐛 Fix context, docstring & prefab properties - Move Sofa.Prefab in its own module in Sofa package - ADD doc for Sofa.Prefab - IMPROVE error handling for parent / parents protected keywords - FIX docstring parameter from __doc__ and prefab params
